feature: spliting in modules
