+++ WithOnlyHibernate/src/main/java/com/infnet/domain/ItemPedidoPK.java
@@ -1,0 +1,28 @@
+package com.infnet.domain;
+
+import lombok.AllArgsConstructor;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import javax.persistence.Embeddable;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import java.io.Serializable;
+
+@Embeddable
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Setter
+@EqualsAndHashCode
+public class ItemPedidoPK implements Serializable {
+    private static final long serialVersionUID = 2087296238312550433L;
+    @ManyToOne
+    @JoinColumn(name = "pedido_id")
+    private Pedido pedido;
+    @ManyToOne
+    @JoinColumn(name = "produto_id")
+    private Produto produto;
+}
