+++ WithSpringDataJPA/src/main/java/com/infnet/Main.java
@@ -1,0 +1,7 @@
+package com.infnet;
+
+public class Main {
+    public static void main(String[] args) {
+        System.out.println("Hello world!");
+    }
+}
