+++ WithSpringDataJPA/src/main/java/com/infnet/domain/ItemPedido.java
@@ -1,0 +1,47 @@
+package com.infnet.domain;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import lombok.AllArgsConstructor;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import javax.persistence.EmbeddedId;
+import javax.persistence.Entity;
+import java.io.Serializable;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Setter
+@EqualsAndHashCode
+@Entity
+public class ItemPedido implements Serializable {
+
+    @JsonIgnore
+    @EmbeddedId
+    private ItemPedidoPK id = new ItemPedidoPK();
+
+    private Double desconto;
+    private Integer quantidade;
+    private Double preco;
+
+
+    public ItemPedido(Pedido pedido, Produto produto, Double desconto, Integer quantidade, Double preco) {
+        id.setPedido(pedido);
+        id.setProduto(produto);
+        this.desconto = desconto;
+        this.quantidade = quantidade;
+        this.preco = preco;
+    }
+
+    @JsonIgnore
+    public Pedido getPedido() {
+        return id.getPedido();
+    }
+
+    public Produto getProduto() {
+        return id.getProduto();
+    }
+}
