+++ WithSpringDataJPA/src/main/java/com/infnet/domain/Pagamento.java
@@ -1,0 +1,47 @@
+package com.infnet.domain;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.infnet.persistenciajava.domain.enums.EstadoPagamento;
+import lombok.AllArgsConstructor;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Inheritance;
+import javax.persistence.InheritanceType;
+import javax.persistence.JoinColumn;
+import javax.persistence.MapsId;
+import javax.persistence.OneToOne;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Setter
+@EqualsAndHashCode
+@Entity
+@Inheritance(strategy = InheritanceType.JOINED)
+public abstract class Pagamento {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer id;
+    private Integer estado;
+
+    @JsonIgnore
+    @JoinColumn(name = "pedido_id")
+    @OneToOne
+    @MapsId
+    private Pedido pedido;
+
+
+    public Pagamento(Integer id, EstadoPagamento estado, Pedido pedido) {
+        super();
+        this.id = id;
+        this.estado = (estado == null) ? null : estado.getCodigo();
+        this.pedido = pedido;
+    }
+}
