+++ WithSpringDataJPA/src/main/java/com/infnet/domain/PagamentoComCartao.java
@@ -1,0 +1,24 @@
+package com.infnet.domain;
+
+import com.infnet.persistenciajava.domain.enums.EstadoPagamento;
+import lombok.AllArgsConstructor;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Setter
+@EqualsAndHashCode
+public class PagamentoComCartao extends Pagamento {
+    private Integer numeroDeParcelas;
+
+    public PagamentoComCartao(Integer id, EstadoPagamento estado, Pedido pedido, Integer numeroDeParcelas) {
+        super(id, estado, pedido);
+        this.numeroDeParcelas = numeroDeParcelas;
+    }
+
+
+}
