+++ WithSpringDataJPA/src/main/java/com/infnet/domain/enums/EstadoPagamento.java
@@ -1,0 +1,36 @@
+package com.infnet.domain.enums;
+
+public enum EstadoPagamento {
+    PENDENTE(1, "Pendente"), QUITADO(2, "Quitado"), CANCELADO(3, "Cancelado");
+
+    private int codigo;
+    private String descricao;
+
+    private EstadoPagamento(int codigo, String descricao) {
+        this.codigo = codigo;
+        this.descricao = descricao;
+    }
+
+    public int getCodigo() {
+        return codigo;
+    }
+
+    public String getDescricao() {
+        return descricao;
+    }
+
+    public static EstadoPagamento toEnum(Integer cod) {
+        if (cod == null) {
+            return null;
+        }
+
+        for (EstadoPagamento estado : EstadoPagamento.values()) {
+            if (cod.equals(estado.getCodigo())) {
+                return estado;
+            }
+        }
+
+        throw new IllegalArgumentException("Id inválido: " + cod);
+    }
+
+}
