+++ docker-compose.yml
@@ -1,15 +1,21 @@
 version: '3.3'
 services:
-  mysql:
+  db:
     image: mysql:8.0.30
     restart: always
     environment:
-      - MYSQL_DATABASE=mysql-persistence
-      - MYSQL_ALLOW_EMPTY_PASSWORD=true
+      MYSQL_DATABASE: 'db'
+      MYSQL_USER: 'user'
+      MYSQL_PASSWORD: 'password'
+      # Password for root access
+      MYSQL_ROOT_PASSWORD: 'password'
     ports:
       - '3306:3306'
+    expose:
+      - '3306'
     volumes:
-      - db:/var/lib/mysql
-      - ./db/init.sql:/docker-entrypoint-initdb.d/init.sql
+      - my-db:/var/lib/mysql
 volumes:
-  db:
+  my-db:
+
+
