--- src/main/java/com/infnet/persistenciajava/PersistenciaJavaApplication.java
@@ -1,13 +1,0 @@
-package com.infnet.persistenciajava;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PersistenciaJavaApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(PersistenciaJavaApplication.class, args);
-	}
-
-}
