--- src/main/java/com/infnet/persistenciajava/domain/Categoria.java
@@ -1,19 +1,0 @@
-package com.infnet.persistenciajava.domain;
-
-import lombok.Data;
-
-import javax.persistence.*;
-import java.util.ArrayList;
-import java.util.List;
-
-@Data
-@Entity
-public class Categoria {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-    private String nome;
-    @ManyToMany(mappedBy = "categorias")
-    private List<Produto> produtos = new ArrayList<>();
-
-}
