--- src/main/java/com/infnet/persistenciajava/domain/Cidade.java
@@ -1,17 +1,0 @@
-package com.infnet.persistenciajava.domain;
-
-import lombok.Data;
-
-import javax.persistence.*;
-
-@Data
-@Entity
-public class Cidade {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-    private String nome;
-    @ManyToOne
-    @JoinColumn(name = "estado_id")
-    private Estado estado;
-}
