--- src/main/java/com/infnet/persistenciajava/domain/Cliente.java
@@ -1,33 +1,0 @@
-package com.infnet.persistenciajava.domain;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.infnet.persistenciajava.domain.enums.TipoCliente;
-import lombok.Data;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.OneToMany;
-import java.util.ArrayList;
-import java.util.List;
-
-@Data
-@Entity
-public class Cliente {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-    private String nome;
-    private String email;
-    private String cpfOuCnpj;
-    private TipoCliente tipo;
-
-    @OneToMany(mappedBy = "cliente")
-    private List<Pedido> pedidos = new ArrayList<>();
-
-    @JsonIgnore
-    @OneToMany(mappedBy = "cliente")
-    private List<Endereco> enderecos = new ArrayList<>();
-}
