--- src/main/java/com/infnet/persistenciajava/domain/Estado.java
@@ -1,21 +1,0 @@
-package com.infnet.persistenciajava.domain;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import lombok.Data;
-
-import javax.persistence.*;
-import java.util.ArrayList;
-import java.util.List;
-
-@Data
-@Entity
-public class Estado {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-    private String nome;
-
-    @JsonIgnore
-    @OneToMany(mappedBy = "estado")
-    private List<Cidade> cidades = new ArrayList<>();
-}
