--- src/main/java/com/infnet/persistenciajava/domain/ItemPedido.java
@@ -1,41 +1,0 @@
-package com.infnet.persistenciajava.domain;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import lombok.Data;
-
-import javax.persistence.EmbeddedId;
-import javax.persistence.Entity;
-import java.io.Serializable;
-
-@Data
-@Entity
-public class ItemPedido implements Serializable {
-
-    @JsonIgnore
-    @EmbeddedId
-    private ItemPedidoPK id = new ItemPedidoPK();
-
-    private Double desconto;
-    private Integer quantidade;
-    private Double preco;
-
-    public ItemPedido() {
-    }
-
-    public ItemPedido(Pedido pedido, Produto produto, Double desconto, Integer quantidade, Double preco) {
-        id.setPedido(pedido);
-        id.setProduto(produto);
-        this.desconto = desconto;
-        this.quantidade = quantidade;
-        this.preco = preco;
-    }
-
-    @JsonIgnore
-    public Pedido getPedido() {
-        return id.getPedido();
-    }
-
-    public Produto getProduto() {
-        return id.getProduto();
-    }
-}
