--- src/main/java/com/infnet/persistenciajava/domain/ItemPedidoPK.java
@@ -1,22 +1,0 @@
-package com.infnet.persistenciajava.domain;
-
-import lombok.Getter;
-import lombok.Setter;
-
-import javax.persistence.Embeddable;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import java.io.Serializable;
-
-@Embeddable
-@Setter
-@Getter
-public class ItemPedidoPK implements Serializable {
-    private static final long serialVersionUID = 2087296238312550433L;
-    @ManyToOne
-    @JoinColumn(name = "pedido_id")
-    private Pedido pedido;
-    @ManyToOne
-    @JoinColumn(name = "produto_id")
-    private Produto produto;
-}
