--- src/main/java/com/infnet/persistenciajava/domain/PagamentoComCartao.java
@@ -1,21 +1,0 @@
-package com.infnet.persistenciajava.domain;
-
-import com.infnet.persistenciajava.domain.enums.EstadoPagamento;
-
-import javax.persistence.Entity;
-
-@Entity
-public class PagamentoComCartao extends Pagamento{
-    private Integer numeroDeParcelas;
-
-    public PagamentoComCartao() {
-
-    }
-
-    public PagamentoComCartao(Integer id, EstadoPagamento estado, Pedido pedido, Integer numeroDeParcelas) {
-        super(id, estado, pedido);
-        this.numeroDeParcelas = numeroDeParcelas;
-    }
-
-
-}
