--- src/main/java/com/infnet/persistenciajava/domain/Pedido.java
@@ -1,34 +1,0 @@
-package com.infnet.persistenciajava.domain;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import lombok.Data;
-
-import javax.persistence.*;
-import java.time.LocalDateTime;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
-
-@Data
-@Entity
-public class Pedido {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    @JsonFormat(pattern = "dd/MM/yyyy HH:mm")
-    private LocalDateTime instante;
-    @OneToOne(cascade = CascadeType.ALL, mappedBy = "pedido")
-    private Pagamento pagamento;
-
-    @ManyToOne
-    @JoinColumn(name = "cliente_id")
-    private Cliente cliente;
-
-    @ManyToOne
-    @JoinColumn(name = "endereco_de_entrega_id")
-    private Endereco enderecoDeEntrega;
-
-
-    @OneToMany(mappedBy = "id.pedido")
-    private Set<ItemPedido> itens = new HashSet<>();
-}
