--- src/main/java/com/infnet/persistenciajava/domain/Produto.java
@@ -1,45 +1,0 @@
-package com.infnet.persistenciajava.domain;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import lombok.Data;
-
-import javax.persistence.*;
-import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
-
-@Data
-@Entity
-public class Produto {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-    private String nome;
-    private Double preco;
-
-    @JsonIgnore
-    @OneToMany(mappedBy = "id.produto")
-    private Set<ItemPedido> itens = new HashSet<>();
-    @JsonIgnore
-    @ManyToMany
-    @JoinTable(
-            name = "PRODUTO_CATEGORIA",
-            joinColumns = @JoinColumn(name = "produto_id"),
-            inverseJoinColumns = @JoinColumn(name = "categoria_id")
-    )
-    private List<Categoria> categorias = new ArrayList<>();
-
-    @JsonIgnore
-    public List<Pedido> getPedidos(){
-        List<Pedido> lista = new ArrayList<>();
-        for (ItemPedido x: itens) {
-            lista.add(x.getPedido());
-        }
-        return lista;
-    }
-
-
-}
