--- src/main/java/com/infnet/persistenciajava/domain/enums/EstadoPagamento.java
@@ -1,36 +1,0 @@
-package com.infnet.persistenciajava.domain.enums;
-
-public enum EstadoPagamento {
-    PENDENTE(1, "Pendente"), QUITADO(2, "Quitado"), CANCELADO(3, "Cancelado");
-
-    private int codigo;
-    private String descricao;
-
-    private EstadoPagamento(int codigo, String descricao) {
-        this.codigo = codigo;
-        this.descricao = descricao;
-    }
-
-    public int getCodigo() {
-        return codigo;
-    }
-
-    public String getDescricao() {
-        return descricao;
-    }
-
-    public static EstadoPagamento toEnum(Integer cod) {
-        if (cod == null) {
-            return null;
-        }
-
-        for (EstadoPagamento estado : EstadoPagamento.values()) {
-            if (cod.equals(estado.getCodigo())) {
-                return estado;
-            }
-        }
-
-        throw new IllegalArgumentException("Id inválido: " + cod);
-    }
-
-}
