--- src/main/resources/application.yml
@@ -1,18 +1,0 @@
-spring:
-  application:
-    name: app
-  jpa:
-    show-sql: true
-    database: mysql
-    hibernate:
-      ddl-auto: update
-    database-platform: org.hibernate.dialect.MySQL8Dialect
-  datasource:
-    url: jdbc:mysql://localhost:3306/db?createDatabaseIfNotExist=true&autoReconnect=true&useSSL=false
-    password: root
-    username: root
-    driver-class-name: com.mysql.cj.jdbc.Driver
-  data:
-    jpa:
-      repositories:
-        enabled: true
