--- src/test/java/com/infnet/persistenciajava/PersistenciaJavaApplicationTests.java
@@ -1,13 +1,0 @@
-package com.infnet.persistenciajava;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class PersistenciaJavaApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
